+++ ui/explore/src/components/PrometheusMetricsFinder/types.ts
@@ -1,0 +1,4 @@
+export interface LabelFilter {
+  label: string;
+  labelValues: string[];
+}
